Refactor AkkaMoneySpec to separate out matchers
