+++ money-akka/src/test/scala/com/comcast/money/akka/TestStreams.scala
@@ -23,7 +23,7 @@ import akka.stream.scaladsl.GraphDSL.Implicits.PortOps
 import akka.stream.scaladsl.{Flow, GraphDSL, Partition, RunnableGraph, Sink, Source}
 import akka.stream.stage.GraphStage
 import com.comcast.money.akka.stream.AsyncUnorderedFlowTracing.TracedFlowOps
-import com.comcast.money.akka.stream.DefaultSpanKeyCreators._
+import com.comcast.money.akka.stream.DefaultStreamSpanKeyCreators._
 import com.comcast.money.akka.stream._
 
 import scala.concurrent.{ExecutionContext, Future}
