+++ money-akka/src/test/scala/com/comcast/money/akka/acceptance/stream/StreamTracingDSLSpec.scala
@@ -19,6 +19,7 @@ package com.comcast.money.akka.acceptance.stream
 import akka.stream._
 import akka.stream.scaladsl.{Flow, Sink, Source}
 import com.comcast.money.akka.Blocking.RichFuture
+import com.comcast.money.akka.SpanHandlerMatchers.{haveSomeSpanNames, haveSomeSpanNamesInNoParticularOrder}
 import com.comcast.money.akka._
 import com.comcast.money.akka.stream._
