+++ money-akka/src/test/scala/com/comcast/money/akka/acceptance/stream/TracedBuilderSpec.scala
@@ -16,12 +16,13 @@
 
 package com.comcast.money.akka.acceptance.stream
 
-import akka.stream.{ Attributes, UniformFanInShape, UniformFanOutShape }
-import akka.stream.scaladsl.{ Balance, Broadcast, Interleave, Merge, MergePrioritized }
-import akka.stream.stage.{ GraphStage, GraphStageLogic }
+import akka.stream.scaladsl.{Balance, Broadcast, Merge, MergePrioritized}
+import akka.stream.stage.{GraphStage, GraphStageLogic}
+import akka.stream.{Attributes, UniformFanInShape, UniformFanOutShape}
 import com.comcast.money.akka.Blocking.RichFuture
-import com.comcast.money.akka.stream.{ UnsupportedUniformFanInShape, UnsupportedUniformFanOutShape }
-import com.comcast.money.akka.{ AkkaMoneyScope, TestStreams }
+import com.comcast.money.akka.SpanHandlerMatchers.haveSomeSpanNamesInNoParticularOrder
+import com.comcast.money.akka.stream.{UnsupportedUniformFanInShape, UnsupportedUniformFanOutShape}
+import com.comcast.money.akka.{AkkaMoneyScope, TestStreams}
 
 class TracedBuilderSpec extends AkkaMoneyScope {
